Cleaning up Real<> for review.
